+++ tests/Jroc.Test262.Tests/built-ins/Promise/resolve/ExecutionTests.cs
@@ -21,4 +21,16 @@ public class ExecutionTests : DiskExecutionTestsBase
     [Fact(DisplayName = "S25.4.4.5_A1.1_T1")]
     public Task S25_4_4_5_A1_1_T1()
         => ExecutionTestFromFile("S25.4.4.5_A1.1_T1");
+
+    [Fact(DisplayName = "arg-non-thenable.js")]
+    public Task arg_non_thenable()
+        => ExecutionTestFromFile("arg-non-thenable");
+
+    [Fact(DisplayName = "resolve-non-obj.js")]
+    public Task resolve_non_obj()
+        => ExecutionTestFromFile("resolve-non-obj");
+
+    [Fact(DisplayName = "resolve-non-thenable.js")]
+    public Task resolve_non_thenable()
+        => ExecutionTestFromFile("resolve-non-thenable");
 }